Move MasterAuditContext to a separate file.
