+++ core/server/master/src/main/java/alluxio/master/file/DefaultFileSystemMaster.java
@@ -43,6 +43,7 @@ import alluxio.master.audit.AsyncUserAccessAuditLogWriter;
 import alluxio.master.audit.AuditContext;
 import alluxio.master.block.BlockId;
 import alluxio.master.block.BlockMaster;
+import alluxio.master.file.MasterAuditContext;
 import alluxio.master.file.async.AsyncPersistHandler;
 import alluxio.master.file.meta.FileSystemMasterView;
 import alluxio.master.file.meta.Inode;
@@ -3287,148 +3288,6 @@ public final class DefaultFileSystemMaster extends AbstractMaster implements Fil
     }
 
     private Metrics() {} // prevent instantiation
-  }
-
-  /**
-   * An audit context for file system master.
-   */
-  @NotThreadSafe
-  public final class MasterAuditContext implements AuditContext {
-    private final AsyncUserAccessAuditLogWriter mAsyncAuditLogWriter;
-    private boolean mAllowed;
-    private boolean mSucceeded;
-    private String mCommand;
-    private AlluxioURI mSrcPath;
-    private AlluxioURI mDstPath;
-    private String mUgi;
-    private AuthType mAuthType;
-    private String mIp;
-    private Inode mSrcInode;
-
-    @Override
-    public MasterAuditContext setAllowed(boolean allowed) {
-      mAllowed = allowed;
-      return this;
-    }
-
-    @Override
-    public MasterAuditContext setSucceeded(boolean succeeded) {
-      mSucceeded = succeeded;
-      return this;
-    }
-
-    /**
-     * Sets mCommand field.
-     *
-     * @param command the command associated with this {@link alluxio.master.Master}
-     * @return this {@link AuditContext} instance
-     */
-    public MasterAuditContext setCommand(String command) {
-      mCommand = command;
-      return this;
-    }
-
-    /**
-     * Sets mSrcPath field.
-     *
-     * @param srcPath the source path of the command
-     * @return this {@link AuditContext} instance
-     */
-    public MasterAuditContext setSrcPath(AlluxioURI srcPath) {
-      mSrcPath = srcPath;
-      return this;
-    }
-
-    /**
-     * Sets mDstPath field.
-     *
-     * @param dstPath the destination path of the command
-     * @return this {@link AuditContext} instance
-     */
-    public MasterAuditContext setDstPath(AlluxioURI dstPath) {
-      mDstPath = dstPath;
-      return this;
-    }
-
-    /**
-     * Sets mUgi field.
-     *
-     * @param ugi the client user name of the authenticated client user of this thread
-     * @return this {@link AuditContext} instance
-     */
-    public MasterAuditContext setUgi(String ugi) {
-      mUgi = ugi;
-      return this;
-    }
-
-    /**
-     * Sets mAuthType field.
-     *
-     * @param authType the authentication type
-     * @return this {@link AuditContext} instance
-     */
-    public MasterAuditContext setAuthType(AuthType authType) {
-      mAuthType = authType;
-      return this;
-    }
-
-    /**
-     * Sets mIp field.
-     *
-     * @param ip the IP of the client
-     * @return this {@link AuditContext} instance
-     */
-    public MasterAuditContext setIp(String ip) {
-      mIp = ip;
-      return this;
-    }
-
-    /**
-     * Sets mSrcInode field.
-     *
-     * @param srcInode the source inode of this operation
-     * @return this {@link AuditContext} instance
-     */
-    public MasterAuditContext setSrcInode(Inode srcInode) {
-      mSrcInode = srcInode;
-      return this;
-    }
-
-    /**
-     * Constructor of {@link MasterAuditContext}.
-     *
-     * @param asyncAuditLogWriter
-     */
-    private MasterAuditContext(AsyncUserAccessAuditLogWriter asyncAuditLogWriter) {
-      mAsyncAuditLogWriter = asyncAuditLogWriter;
-      mAllowed = true;
-    }
-
-    @Override
-    public void close() {
-      if (mAsyncAuditLogWriter == null) {
-        return;
-      }
-      mAsyncAuditLogWriter.append(this);
-    }
-
-    @Override
-    public String toString() {
-      if (mSrcInode != null) {
-        short mode = mSrcInode.getMode();
-        return String.format(
-            "succeeded=%b\tallowed=%b\tugi=%s (AUTH=%s)\tip=%s\tcmd=%s\tsrc=%s\tdst=%s\t"
-                + "perm=%s:%s:%s%s%s",
-            mSucceeded, mAllowed, mUgi, mAuthType, mIp, mCommand, mSrcPath, mDstPath,
-            mSrcInode.getOwner(), mSrcInode.getGroup(),
-            Mode.extractOwnerBits(mode), Mode.extractGroupBits(mode), Mode.extractOtherBits(mode));
-      } else {
-        return String.format(
-            "succeeded=%b\tallowed=%b\tugi=%s (AUTH=%s)\tip=%s\tcmd=%s\tsrc=%s\tdst=%s\t"
-                + "perm=null",
-            mSucceeded, mAllowed, mUgi, mAuthType, mIp, mCommand, mSrcPath, mDstPath);
-      }
-    }
   }
 
   /**
